test client added

+++ TrackingTableRecord.java
@@ -3,9 +3,9 @@ package TrackingTable;
 public class TrackingTableRecord implements Comparable<TrackingTableRecord> {
 	private Range range;
 	private char statusCode;
-	private long transferCode;
+	private int transferCode;
 
-	public TrackingTableRecord(Range range, char statusCode, long transferCode) {
+	public TrackingTableRecord(Range range, char statusCode, int transferCode) {
 		this.range = range;
 		this.statusCode = statusCode;
 		this.transferCode = transferCode;
@@ -27,11 +27,11 @@ public class TrackingTableRecord implements Comparable<TrackingTableRecord> {
 		this.statusCode = statusCode;
 	}
 
-	public long getTransferCode() {
+	public int getTransferCode() {
 		return transferCode;
 	}
 
-	public void setTransferCode(long transferCode) {
+	public void setTransferCode(int transferCode) {
 		this.transferCode = transferCode;
 	}
